added common ancestor
