+++ PR8/commonAncestor.js
@@ -1,0 +1,24 @@
+import { familyTree } from "../utils.js";
+
+// 1. listen
+// 2. clarify. given a family tree, I need to find the youngest common ancestor for two given aliens.
+
+// example:
+//       5
+//     /  \
+//    3    7
+//   /\    /\
+//  2  4  6  8
+
+// [2,8] = 5
+// [2,4] = 3
+// [6,7] = 5
+
+// 3. function signature
+// function youngestAncestor(node1, node2){
+// operations
+// return youngest common ancester node
+// }
+
+// non-code steps:
+// 1.
