+++ index.html
@@ -9,8 +9,9 @@
     (open the console)
 
     <script src="PR8/longLines.js" type="module"></script>
-    <script src="PR8/multiplyNums.js"></script>
-    <script src="PR8/familyTrees"></script>
+    <script src="PR8/multiplyNums.js" type="module"></script>
+    <script src="PR8/familyTrees.js" type="module"></script>
+    <script src="PR8/commonAncestor.js" type="module"></script>
     <script src="utils.js" type="module"></script>
   </body>
 </html>
